Altered query to retrieve mappings to evaluate
— The query that loads a new synset mapping to be evaluated has been
altered. Now it randomizes the synset being selected to make sure that
positive and negative examples are mixed together. Even with the
mixing, the query will still retrieve all synsets related to a certain
supersense and will only change to another supersense once there are no
more mappings for that supersede to be evaluated